requirements.txt
flask
requests
azure-cognitiveservices-vision-face
azure-cognitiveservices-search-imagesearch
opencv-python
